separate progress report code
