+++ xpano/algorithm/progress.h
@@ -1,0 +1,41 @@
+// SPDX-FileCopyrightText: 2023 Tomas Krupka
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#pragma once
+
+#include <atomic>
+
+namespace xpano::algorithm {
+
+enum class ProgressType {
+  kNone,
+  kLoadingImages,
+  kStitchingPano,
+  kAutoCrop,
+  kDetectingKeypoints,
+  kMatchingImages,
+  kExport,
+  kInpainting,
+};
+
+struct ProgressReport {
+  ProgressType type;
+  int tasks_done;
+  int num_tasks;
+};
+
+class ProgressMonitor {
+ public:
+  void Reset(ProgressType type, int num_tasks);
+  void SetNumTasks(int num_tasks);
+  void SetTaskType(ProgressType type);
+  [[nodiscard]] ProgressReport Progress() const;
+  void NotifyTaskDone();
+
+ private:
+  std::atomic<ProgressType> type_{ProgressType::kNone};
+  std::atomic<int> done_ = 0;
+  std::atomic<int> num_tasks_ = 0;
+};
+
+}  // namespace xpano::algorithm
